+++ libvast/vast/table_slice.hpp
@@ -136,14 +136,15 @@ public:
   /// @param builder The builder to pack *x* into.
   /// @param x The table slice to pack.
   /// @returns The flatbuffer offset in *builder*.
-  friend caf::expected<flatbuffers::Offset<fbs::table_slice_buffer::v0>>
+  friend caf::expected<flatbuffers::Offset<fbs::FlatTableSlice>>
   pack(flatbuffers::FlatBufferBuilder& builder, table_slice_ptr x);
 
   /// Unpacks a table slice from a flatbuffer.
   /// @param x The flatbuffer to unpack.
   /// @param y The target to unpack *x* into.
   /// @returns An error iff the operation fails.
-  friend caf::error unpack(const fbs::table_slice::v0& x, table_slice_ptr& y);
+  friend caf::error
+  unpack(const fbs::table_slice::legacy::v0& x, table_slice_ptr& y);
 
 protected:
   // -- member variables -------------------------------------------------------