+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Every entry has a category for which we use the following visual abbreviations:
 
 ## Unreleased
 
+- ⚡️ The on-disk format for table slices now supports versioning of table slice
+  encodings. This breaking change makes it so that adding further encodings or
+  adding new versions of existing encodings is possible without breaking again
+  in the future. [#1143](https://github.com/tenzir/vast/pull/1143)
+
 - ⚡️ CAF-encoded table slices no longer exist. As such, the option
   `vast.import.batch-encoding` now only supports `arrow` and `msgpack` as
   arguments. [#1142](https://github.com/tenzir/vast/pull/1142)